+++ AdMob/src/main/java/com/mopub/mobileads/GooglePlayServicesBanner.java
@@ -12,6 +12,7 @@ import com.google.android.gms.ads.AdListener;
 import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.AdSize;
 import com.google.android.gms.ads.AdView;
+import com.google.android.gms.ads.LoadAdError;
 import com.google.android.gms.ads.MobileAds;
 import com.google.android.gms.ads.RequestConfiguration;
 import com.mopub.common.LifecycleListener;
@@ -29,6 +30,7 @@ import static com.google.android.gms.ads.RequestConfiguration.TAG_FOR_UNDER_AGE_
 import static com.google.android.gms.ads.RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_TRUE;
 import static com.google.android.gms.ads.RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_UNSPECIFIED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CLICKED;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CUSTOM;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_ATTEMPTED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_FAILED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_SUCCESS;
@@ -181,12 +183,16 @@ public class GooglePlayServicesBanner extends BaseAd {
         }
 
         @Override
-        public void onAdFailedToLoad(int errorCode) {
+        public void onAdFailedToLoad(LoadAdError loadAdError) {
             MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME,
-                    getMoPubErrorCode(errorCode).getIntCode(),
-                    getMoPubErrorCode(errorCode));
+                    getMoPubErrorCode(loadAdError.getCode()).getIntCode(),
+                    getMoPubErrorCode(loadAdError.getCode()));
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Failed to load Google " +
+                    "banners with message: " + loadAdError.getMessage() + ". Caused by: " +
+                    loadAdError.getCause());
+
             if (mLoadListener != null) {
-                mLoadListener.onAdLoadFailed(getMoPubErrorCode(errorCode));
+                mLoadListener.onAdLoadFailed(getMoPubErrorCode(loadAdError.getCode()));
             }
         }