+++ AdMob/CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Changelog
+  * 19.3.0.0
+    * This version of the adapters has been certified with AdMob 19.3.0 and MoPub 5.13.1.
+    * All ad formats: Remove load failure callbacks that return an `int` (deprecated by Google) and use the equivalent (if any) that returns a `LoadAdError`.
+    * Rewarded video: Remove `onRewardedAdFailedToShow(int)` (deprecated by Google) and use `onRewardedAdFailedToShow(AdError)`.
+
   * 19.2.0.1
     * Fix interstitial load listener usage.